Add SectionedHeatExchanger draft and change inheritance structure
