+++ src/tespy/components/__init__.py
@@ -14,6 +14,7 @@ from .heat_exchangers.movingboundary import MovingBoundaryHeatExchanger  # noqa:
 from .heat_exchangers.parabolic_trough import ParabolicTrough  # noqa: F401
 from .heat_exchangers.parallel import ParallelFlowHeatExchanger  # noqa: F401
 from .heat_exchangers.simple import SimpleHeatExchanger  # noqa: F401
+from .heat_exchangers.sectioned import SectionedHeatExchanger  # noqa: F401
 from .heat_exchangers.solar_collector import SolarCollector  # noqa: F401
 from .nodes.droplet_separator import DropletSeparator  # noqa: F401
 from .nodes.drum import Drum  # noqa: F401
